StreamingWorker can now be canceled, also fixed threading concurrency issue.
If the StreamingWorker finished contouring before the REQUEST_STREAMING_UPDATE
call to the representation had occurred we would enter a state where we never
fetch the computed polydata but always have empty output. For now
we state we are never finished to solve the problem, but I expect in the long
term we will need a different solution.