Added new DataPropertiesPanel.
Added new DataPropertiesPanel to show information about the active data
source.

Also fixed issue when data-range for the transformed data was not
updated. The problem was due to missing "Modified" on the data array.
Cleaned that up by adding util.set_scalars() as the proper mechanism to
pass transformed scalar back to the pipeline.