Adding a Recontruct menu option, encode the Python file
The Python reconstruction file, supplied by Robert Hovden, uses the
numpy routines exposed in tomviz to reconstruct a volume. The file
is statically encoded right now, it could be modified to read this
from the installed files if that was preferable. Some variables at
the top need editing at present to supply some information not yet
encoded in the data structures available to tomviz.