Add ability to apply alignment, insert new data
This adds some code to apply the offsets specified, and add them to
the pipeline as a new dataset. Still need to add the capability to
actually save the offsets out too, and possibly to set the fill
value. There are also a number of features that we may want for the
alignment user interface.