Redid the histogram creation thread
The histogram will now check when the data is ready if the same dataset
is still active and only display the new histogram if it is.  It also
respects Qt's rules for thread ownership of the vtkSmartPointer objects
and will no longer segfault when tomviz is closed and a histogram is in
progress.