Do not add actor when transparent -- fixes #157
Having a transparent (alpha=0) actor in the scene changes the way that volume
rendering interacts with the polygonal geometry.  When the translucent actor is
present, the volume rendering is layered on top of the polygonal geometry
regardless of depth.  This change refuses to add an actor from the scene if it
will always be transparent.