Make the data/module properties context aware
It is confusing to see the data properties and module properties when
only one of the two can be selected at any given time. Added some new
signals to indicate when a given item is active, and some logic to the
main window to switch in the appropriate panel.

This fixes issue #46.