Refactor edit operator dialog to use operator-specified widget
This commit adds a method to the tomviz::Operator class to get the
editor widget for that type of operator.  This is displayed in a dialog
when the operator is double-clicked in the operators list.  The widget
must be a subclass of QWidget that understands how to apply the changes
to the operator that is being edited in a virtual Qt slot.  The rest of
this commit is the changes necessary for these changes to compile and
run, although the crop operator does not support the new widget-based
edit dialog yet: it just has a stub for the getEditorContents() method.