EditOperatorDialog: optionally add to data source on first apply
This commit adds an optional DataSource argument to the
EditOperatorDialog that when non-NULL causes the dialog to add its
operator to that data source when Apply or OK is clicked for the first
time.  This removes the need to externally add callbacks to listen for
this first event and add the operator in all the 'add an operator' menu
item reaction classes.