Add UI for rotation align
This adds the UI part of the rotation axis alignment for tilt series
datasets.  The UI is a dialog that shows slices of the dataset along
with the axis of rotation (movable) and three views at the bottom that
currently show sinograms of the dataset at user-specified slices.  This
UI can be accessed via a menu item in the Tomography menu