Call QObject::disconnect in right direction
Every time these functions were called it was adding the connection
again resulting in that many more calls to setDataSource the next time
a signal fired.  QObject::disconnect has to be called on the source
object and is given the object that is receiving the signal as its
argument.  We had the calls backwards.