Set up contour module for coloring by label map
Normally, this would be simple to implement by just changing the color
array name. However, the FlyingEdges filter does not interpolate
scalar arrays to keep things fast. Hence, to get the label map values
on the contour, we need to use the Probe filter to sample the scalar
values onto the contour surface.

Changed the logic of this class to modify the pipeline as needed to
get the label map values interpolated on the surface. The logic
preserves the speed of FlyingEdges when coloring by label map is not
requested.