Refactoring of color/opacity editor
This commit wraps the color/opacity editor functionality into a single
class, vtkChartHistogramColorOpacityEditor.  Having all the
functionality in one class simplifies the CentralWidget code
somewhat. This change also makes it so that only one QVTKWidget is
needed for the color/opacity editor. The previous functionality is all
as it was before.

One improvement is that the color bar left edge is aligned with the
left edge of the histogram.