Delay the update of the data properties panel
There is a strange interaction between Qt 5 and QVTKWidget-based classes
in docks, such as the color transfer function editor. If no one ever
looks at this panel it is nice to avoid doing all this work, and by
doing so we delay initialization until the widget is visible.

This is a pretty hackish solution to that bug, and it needs more testing
on other platforms to see if it helps everywhere. We also need to do
more work to ensure the VTK-based OpenGL widgets always initialize their
context successfully, but we can do that later.