Give operators midpoint of pipeline image when editing them
When an operator is edited we occasionally need information about the
dataset to construct its editor dialog.  For instance on the Crop
operator we need the origin, spacing and extent of the image to position
the widget and determine the range of valid entries in the input boxes.
But these values can be changed by subsequent transforms in the
pipeline, so the "current" values for that DataSource may not be valid.
This gets even stranger when the editor widget displays the
"intermediate" data like the translation align widget (Reported in #434)

This commit adds a method which gets a copy of the data in a DataSource
at a particular point in its data transform pipeline.  This new data
object is then passed to the Operator in getEditorWidget() so that the
editor widget will have accurate data for the input of the Operator.