Added custom pipeline model and view classes
These classes use the more modern Qt APIs for managing data of this
type, and expose the operators as an integral part of the pipeline.
This carries forward most functionality, such as keyboard shortcuts,
context menu, icons, labels, etc. It also conveys more directly what
is happening to the data from source through to display/rendering.

The internal tree model attempts to enforce a hierarchy on the model.
The data source has children which are either modules or operators,
when new modules or operators are added then they are reordered to
make all modules the children of the final operator. This gives a
visually simple representation where a data source flows to zero or
more operators, flows through each one, and is then displayed by the
modules in the view. Switching views toggles whether modules are
enabled or not.