Compute some label object shape attributes
Introduces the use of ITK to compute label object shape attributes and
store them in the field data of the output data set. A label object
consists of all the voxels that have a particular label
value. Typically, these voxels are connected in some way, such as in a
connected component.

The attributes computed are object surface area (perimeter in ITK
parlance), object volume (number of voxels multiplied by the voxel
volume), and surface area to volume ratio.

Added to the Connected Components module.