Solve some invalid model index errors in PipelineModel
Fixes some QModelIndex errors reported by
PipelineModel::endInsertRows() when a data source is first loaded. The
problem was that the last row was not the same as the first when only
one new row is being inserted, but it should be.

Some invalid model index errors persist when an operator is first added.