Line up Operators and Modules in Pipelines view
In preparation for adding derived data sets, unindent the Modules so
that users do not confuse Modules with derived data sets. Operators
are still on top and Modules are on the bottom. They red X's on the
right of the Operators should distinguish them sufficiently from
Modules with eyeball visibility icons.