Modify BinaryThreshold python to return label map object
It was setting a label map array in the original data set, but now it
returns a label map object in a Python dictionary of results.  The
OperatorPython class registers this data in the proxy manager and adds
a DataSource that wraps it as a child DataSource in the Operator.

Also added a BinaryThreshold JSON description file. It is not used
yet, but will be.