Added flag indicating whether an Operator has a child
We need a flag that indicates whether an Operator has a child
dataset. The JSON will set this up, and we'll know when to look for a
child data set when processing the output of the Operator. It is also
needed to set up the UI prior to running the Operator.