Reverse connected components labeling
Flip the labels so that the largest component has the largest label
value and the smallest component has the smallest value. The label
value is really arbitrary, so changing the labeling in this way
shouldn't be a big deal. It does, however, make display of the largest
components via color map nicer because the largest component now has a
value most different from the background label.