Replace PyErr_Print() with PyErr_PrintEx(0)
Use PyErr_PrintEx(0) to prevent sys.last_traceback being set.
sys.last_traceback holds a reference to any parameters
passed to a PyObject_Call. This can cause a temporary "leak"
until sys.last_traceback is reset. This can be a problem if
the object in question is a VTK object that holds a reference
to a large memory allocation. The call the PyErr_Clear() is
also unecessary as the PyErr_PrintEx already resets the error
indicator.

See http://bugs.python.org/issue1756389 for more details on this
issue.