Added Otsu multiple threshold operator
This operator exposes ITK's OtsuMultipleThresholdsImageFilter in
Tomviz. One input parameter is expected, the number of thresholds to
find. The number of labels in the output image is the number of
thresholds plus one.

The threshold values determined by the operator are reported on the
command line and Python console. Some UI infrastructure will be
required to report these values back to the UI and enable their
editing.