Expose some additional volume rendering features
This commit switches from the ParaView volume renderer to a locally
constructed one. This gives full access to the volume renderer,
properties, and mapper. So far I have just exposed a couple of the
features, and plan to add more in future commits.

I still need to add support for serialize/deserialize. The proposed
changes are being tracked in issue #633.