Canceled operators should not be removed from pipeline
The assumption here was that if applyTransform return false then
the operator was canceled, this is not the case it may mean an
error has occured. Either way we don't want to remove the operator
from the pipeline. We may want to change the applyTransform method
to return an enum to differentiate between cancel and error.