Remove Modules from child data sources
Modules created to visualize child DataSources were not being removed
when the Operator that created the child DataSource was removed, so
the Module visualizations would persist in the render view with no way
to modify or remove them. Fixed that by removing child DataSource
Modules while the Operator is being removed in the PipelineModel.

Fixes #624