Synchronize VTK object values with its proxy properties
When serializing the program state, the ScalarOpacityFunction proxy's
Values property was not up-to-date with the VTK object it represented.
ParaView property values usually are set on the property and then
pushed to the VTK object. In the case of the ScalarOpacityFunction,
editing of values is done directly on the vtkPiecewiseFunction and
the values never make it back to the ParaView-side property.

Solved for now by manually copying the VTK-side values to the
ParaView property prior to serialization.