Fix issue #654 - check index is valid before deletion
For some reason the packaged OX X binary has more issues with additional
events being generated, in this case several delete key events were
generated after the initial one with invalid indexes. This adds some
checking in a few places to ensure we only return items for a valid
index, and that the deletion is only requested for a valid index.