Don't automatically change minimum image extent when setting array
Setting the minimum extent to (0, 0, 0) in utils.set_array(...)  is
not always the right thing to do, especially after an Operator changes
the size of the image.

Now, tomviz.utils.set_array() takes an optional argument that enables
setting of a new minimum image extent. If this argument is not set,
the minimum of the extent is left alone. If it is set, the minimum
extent is updated in the data object. In all cases, the maximum extent
is updated so that the data object dimensions are the same as the
shape of the array that is being set.