Prevent QJsonObject from going out of scope
childDatasetArray[0].toObject()["name"] results in a QJsonValueRef
to a QJsonObject that has gone out of scope.
childDatasetArray[0].toObject() returns a QJsonObject object, we
then take a reference to one of the properties the underlying QJsonObject
object then goes out of scope and we are left with references that
refer to a dangling pointer.