Handle cases where data is added when spreadsheet view is active
There was an assumption in LoadDataReaction and
PythonGeneratedDatasetReaction that the current active view is a
Render View. When that isn't the case, visualization modules are not
added correctly, signals and slots are not set up properly, and a
crash might occur. In any case, Tomviz becomes unstable.

Fixed this problem by checking that the active view is a Render View,
and if not, look for one. If there isn't one, create a Render View.

Fixes #587