Fix too many SpreadSheetViews being opened
If a SpreadSheetView was not active when double-clicking a table
OperatorResult in the PipelineView, a new SpreadSheetView would be
created, leading to potentially many SpreadSheetViews being created to
show the same data.

Instead, search for an existing SpreadSheetView and show the table
data there if possible. Also, make the selected SpreadSheetView the
active one to ensure it is highlighted for the user.

Fixes #628