Gracefully handle pixel types unsupported by ITK wrapping
Perform automatic casting of a VTK image data to a suitable pixel type
prior to conversion to an ITK image. For each supported VTK pixel
type, look through a list of supported pixel types in ITK that is
ordered by preference. Preference is based on the suitability of the
destination type to represent the source type while minimizing the
destination image size.

Fixes #635