Add utility function to get Python type for VTK, ITK data types
ITK filter member functions sometimes accept only certain Python
numeric types. For example, an ITK filter templated for 'unsigned
char' voxels may have a member function that requires an 'int'
argument from Python - no casting appears to be done behind the scenes
to handle float types, for instance. As a result, an argument passed
in as a 'float' would need to be converted to an 'int' before passing
it to this member function.  Otherwise, an exception may occur.

The utility function in this commit returns a suitable Python numeric
type when given a vtkImageData or an ITK image. This type can be used
to handle casting in the Python script.