Make Connected Components operate directly on label images
The initial implementation of the Connected Components operator was
written before label maps were created as child datasets by operators
such as Binary Threshold. Because of this, it first performed a binary
threshold to obtain a label map prior to running the ITK Connected
Components filter. Changed the implementation to remove the threshold
step so that the operator can be applied to label maps generated by
any operator and to avoid a likely redundant thresholding operation.