Force integral types for Otsu threshold output labels
Limitations in the Otsu threshold filter wrapping meant that the
output of the Ostu threshold filter had to be the same as the input.
For input images of type float or double, this produced output
labelmap datasets with float or double labels, respectively. Filters
such as connected components would not work with float or double
labels, so we now cast the output of the Ostu threshold filter to an
integral type (unsigned char) if it otherwise is not.