Update the icon naming for retinal displays
This change follows the recommended naming convention for retinal
display support in Qt 5.x. The normal 32x32 icons have no suffix, and
the double sized 64x64 have an @2x suffix so that they are substituted
when the device pixel ratio is 2. This should result in no change on
most platforms, and does not require a retinal display as it will fall
back to the original icons in all other places.