Make BinaryThreshold cancelable and report its progress
This shows how to make ITK-based data transforms cancelable and how to
report their progress. We add a progress observer command that invokes
a local function each time a progress event is emitted from the ITK
filter. This function relays the progress to Tomviz and checks to see
if the ITK filter should be aborted. If canceled, the filter throws a
RuntimeError that we catch and ignore, then return from the
transform_scalars function early.

This example also splits the transform into different steps, each of
which is assigned a somewhat arbitrary progress percentage that is
completed when the step is finished. Status messages are set
throughout.