Made it possible for a palette to override custom background
Changed a ParaView-ism where setting a custom background color unlinks
the background property from the global color palette background
property. Now the link is re-established if it is not present whenever
a color palette is applied. This has the effect of applying the color
palette background to the view background.