Add a Segment Particles transform
This segments the particles in the LARGE_PtCu_NanoParticles dataset. The
segmented particles have all pores filled in.

The only parameter of the transform is the minimum particle size. This
is specified as a radius in voxels. Structures smaller than this size is
suppressed. This algorithm enhances spherical structures in the image
and suppress streak artifacts.