Add Segment Pores transform
This does some pre-processing on the input image, then does automatic
thresholding to find the basis for the pores. Next, the particles are
identified and encapsulated to distinguish internal pores and even pores
that are exposed to the surface. A morphological watershed filter applied to
the encapsulated pores distance map separates pores based on the minimum
radius. Thresholded components smaller than the minimum radius are also
removed.