Hide GradientOpacityWidget depending on the active module.
GradientOpacityWidget is hidden if the active module does not support
this type of transfer function (currently only ModuleVolume supports it).
A button was added in HistogramWidget to toggle the visibility of
GradientOpacityWidget, this button is also disabled for unsupported modules.

Added a checkbox in ModuleVolume to enable/disable gradient opacity function.

Some fixes for the chart layout, scale, axis and histogram visibility.

Cleanup and clang format.