Add capability to scan for the data node in EMDs
The path to the data node is not fixed, added some code that scans
for the location. This still has a couple of issues with the sameples
I have from NCEM but it is now able to read those sample files. It
also round trips successfully. Still need to add support for units.