Basic caching works, seems to behave as expected
This adds a new operator type, and puts it into the context menu for the
pipeline. When it is first executed it creates a child data set with a
deep copy of the output of the operator pipeline. After that it reports
success, and doesn't touch its child data set (or execute anything).