Handle case of only one Operator in pipeline
The logic in DataSource::getCopyOfImagePriorTo(Operator*) was
mishandling the case where only one Operator was in the pipeline. If
there is only one Operator, the copy of the image prior to it is the
original data source. We need only rerun the pipeline if there is more
than one Operator.