Fix volume selection widget bounding box initialization
The volume selection widget bounding box was accidentally collapsed to
a point, making intersection tests with the data bounding box
unreliable. As a consequence, it was almost impossible to get a
reasonable result when any part of the widget bounding box fell
outside the data bounding box. This commit fixes an off-by-one
indexing error.