Add signal when Operator is about to be destroyed
QObject has a destroyed(QObject*) signal, but by the time it is
emitted, the Operator destructor is complete and a QObject pointer for
an Operator can no longer be qobject_cast to an Operator pointer.
Emitting a new signal enables receiving objects to access the Operator
pointer while it is still around.