Keep a map of operators to their editor dialogs
When double-clicking, check to see if an editor dialog already exists
for the selected Operator. If it does, show it and raise it. If not,
create it and add it to the map, then show it. This approach supports
editor dialogs for multiple operators to be open at the same time,
but prevents more than one editor dialog for a single operator
from being opened.

Map entries for Operators that are deleted are removed via a
signal/slot connection.