Added vtkTransferFunctionBoxItem.
This context item is meant to be used to select a region within the
2D histogram and specify a transfer function for it.

Added panning and corner dragging interaction.

Added color/opacity transfer function members and a method to query
the pose of the box. These will be used by the chart to generate the
color mapped transfer function.