Enable saving of OperatorResults
If an OperatorResult, such as the table data produced by Label Object
Attributes, is selected when "File -> Save Data" is run, save that
data rather than the active DataSource. This works by keeping track of
the active OperatorResult in ActiveObjects, then when writing data
writing the OperatorResult if it is non-null, rather than the active
DataSource.