Prevent deletion of child DataSources
It was possible to delete a child DataSource via the context menu or
by pressing delete when the child DataSource was selected in the
PipelineView. This change adds a function to the ModuleManager to
check whether a DataSource is a child DataSource and uses it to
prevent deletion of children only. Deleting a parent of a child
DataSource continues to remove the child.

Note: deletion of child data sources seems to be fairly well supported
in the sense that it doesn't cause crashes, but there are some missing
features in the pipeline interaction, such as a context menu to re-run
an Operator whose child has been deleted. Preventing independent
deletion is likely the simplest interaction improvement for now.