Refactor, less shouty, Windows compile ERROR
I hate all-caps if I am honest, but this is motivated by a strange compiler
error in the AcquisitionWidget.cxx file on Windows. It appears that ERROR is
being replaced with constant in Operator.h. This works around that compiler
issue, and makes me a little happier in the process...