Remove assumption that Scalars are set by reader
It is possible in some cases that a reader will not set the active
Scalars array. For example, when reading a .vti file, it is possible
to not load the array that is designated as the Scalars array in the
data set. In such cases, tomviz would report that loading the data set
failed, even though a data array may have been read.

This commit changes the logic to check whether there are point data
arrays available instead of a designated Scalars array to better
handle such corner cases.