Compute the magnitude of multicomponent datasets
Changing the component number from 0 to -1 tells the scalar range
computation to use the magnitude of each tuple in a multicomponent
dataset rather than a particular tuple component. Prior to this change
the code was always computing the scalar range for the 0th component,
which was fine in single-component datasets, but was only giving the
range for the first component in multicomponent datasets. Changing it
to -1 computes the magnitude instead. For single-component datasets,
setting the component to -1 still yields the actual range (not
magnitude) of the scalars, so it will be accurate for any dataset with
negative values.