Add component option to CalculateHistogram function
Similar to how the scalar range for a particular component in a
multicomponent dataset can be requested in vtkDataArray::GetRange(),
this commit adds a component argument used to request which array
component is to be used when calculating the histogram. A component
value of 0 or more refers to an individual array component, while -1
has the same meaning as vtkDataArray::GetRange() (it computes the
vector magnitude of multicomponent images, or the range of the actual
values if the number of components is 1).