Read back in the units from an EMD file
This currently assumes the dimension array is expressed in 32 bit floats,
this can be generalized in the future if desired. It also takes care of
flipping the spacings back and forth on read/write to express volumes in
the "C" order for compatibility with other programs.

We should discuss whether some of these things might be standardized at
the upcoming meeting, or if there is a need for generality on types used.