Refactor transfer function 2D to store data in DataSource
This creates a transfer function 2D box object that stores the
information necessary to generate a 2D transfer function.  This box
object is stored in the data source/module so that having separate
transfer function 2Ds is supported.  This is necessary for saving the
transfer function 2D which will be done later on this branch.