Consider number of image components in available Modules
Some modules cannot be applied when the DataSource has multicomponent
active scalars. This commit provides a method to query the number of
components in the DataSource, and consults it when returning the list
of applicable Module types. At the moment, the Contour, Volume, and
Threshold Modules are not applicable to multicomponent images.

Also added a static function in ModuleFactory to check whether a
module of a given name is applicable to a DataSource and view.
This function is used to enable/disable module buttons in the
UI as needed.