Refactoring data source, proxy handling
This uses the trivial producer proxy as the default proxy, makes the
function that returns a trivial producer public and removes boilerplate
code that was performing the SafeDownCast in many places, and registers
the producer. It also removes access to the file proxy (not generally
useful), and ensures the color map is correctly initialized.