Add active array concept for Modules
Added new API that lets you set and query an arbitrary number of
"active" point data arrays. The active arrays are indexed. It is up to
the Module to determine how to use each indexed array. In
ModuleOrthogonalSlice, for example, the active array is used to
determine the ColorArrayName in the slice representation.