Compute range from histogram table data
When the "Reset data range" button was clicked, tomviz would delegate
the data range computation to the active ParaView representation.
This isn't reliable as there may be no active representation - not
every Module in tomviz uses a ParaView representation.

Instead, compute the range from the vtkTable representing the data
histogram in the histogram view. This is much more reliable and
directly links the data range to the displayed histogram.