Enable importing of Python operators via menu item
tomviz now always shows the "Custom Transforms" menu, which now
features a new menu item "Import Custom Transform". When selected,
this item brings up a file dialog one can use to select a Python
(*.py) file to import a custom transform. The located Python file will
be copied to <home>/tomviz. If a .json file with the same base file
name is located in the same directory as the selected Python file, it
will also be copied to the <home>/tomviz directory.

Imported custom transforms appear immediately in the "Custom
Transforms" menu