Eliminate duplicate code path
Whenever a generated data source was added, most of the same steps for
adding it to tomviz as when data was loaded from a file were executed
in a duplicate code path. Critically, however, the addition of a
Pipeline to handle the data source was missing. Resolved this by
eliminating the duplicate code and invoking
LoadDataReaction::dataSourceAdded() instead.