Add optional 'name' parameter to set_array* functions
set_array() and set_array_from_itk_image() now have an optional 'name'
parameter that lets you choose the name of the output array.  If not
provided, the name of the current active scalar array will be used. If
provided, the array will be added as an additional data array in the
data set with the given name.