Add progress function to notify tomviz of updated data
This is rudimentary support for letting Python operators notify tomviz
that temporary data results are available.

Envisioned usage:

self.progress.data_updated()

In addition, added a progress attribute for passing data from an
operator to the user interaction thread. Currently not working because
the pybind side is not happy with the VTK-Python-wrapped vtkImageData
it receives.