Added a document to step through passive watching
This adds some detail on how to start the server, connect to it, and how
to inject some files from an existing image stack in the case that you
don't have an electron microscope handy (or would rather experiment
before using time on the instrument). This is very much in development,
and will likely change significantly.