Create child datasets prior to running transforms
Child datasets are declared in the JSON, so we can create them prior
to running an operator. We do this to avoid deadlocking caused by the
Python GIL on the main thread trying to obtain a lock on from the
worker thread that can occur in some of the ParaView Python tracing
code.

In addition, converted the connection notifying the main thread of
data updates to child data to a regular connection rather than a
blocking queued connection to avoid the same Python GIL ensurer
deadlock issue.

Finally, added code to signal to tomviz that the DataSource has
been changed and that a render is needed whenever a child source
data is updated.