Fix a race condition and deadlock
Make creation of the child data source a blocking operation on the
worker thread so that the child data source is ready to go before
running the Python operator code.

Also, move creation of a Python context until after the child data is
created to avoid lock contention between the Python GIL ensurer from
the ParaView Python tracing functions and the Python context.