Re-add capability to specify child dataset in return value
Prior to the live update capabilities that enable intermediate results
to be sent to the main thread for rendering, child datasets were
created in Python and returned in a dictionary return value.  We lost
the ability to do this in the live update work, but this commit
restores the capability.