Rework ownership chain for the pipeline a little
This fixes a number of memory leaks, effectively stop setData in
DataSource from decrementing the reference count, and have the pipeline
objects ensure the data being processed is not deleted. This removes
some asymmetry that was difficult to manage. The copyData method always
returns an object that must be deleted, and these changes keep that
local to the copyData call.