Ensure the default split favors the 3D viz
This fixes #1524 by delaying some of the layout. The widget now saves
and restores its own state. This is delayed as the histogram widget has
a minimum height that is initially too large forcing it to occupy more
space. Using a few persistent settings to indicate first showing of the
widget.