Emit signal that the active scalars may have changed in copyData()
When a vtkImageData copied into the current vtkImageData has different
arrays, this could lead to a warning message from vtkPVImageSliceMapper
(used by ModuleOrthogonalSlice) because the color array used by the
mapper was invalid for the new data. Emitting the signal ensures that
the ColorArrayName is set to something valid in ModuleOrthogonalSlice,
avoiding the warning.