Simplify and speed up histogram calculation
Looking at histogram calculation, especially for single component
volumes, and speeding it up with a number of small changes. Single
component is our primary use case right now, and looking at a 1025^3
volume of unsigned chars the performance was improved.

Typical numbers measured locally:

 * 10.9s - original code
 * 8.8s - remove std::min call from the tight loop
 * 4.9s - remove the isFinite call/branching
 * 3.9s - go from division to multiplication by the inverse

This is currently single-threaded, and so the next logical step would be
to add back in a multithreaded version using VTK-m.