Fix a rare crash on startup due to the acquisition dialog
m_format and m_extension were not initialized explicitly,
but would rely on signals fired by a child widget.
At times, one of these signals wouldn't fire on time,
leaving one of the two uninitialized, and as a consequence a list
that is required to have 5 elements would be empty.
Trying to access its elements, would eventually cause a
segmentation fault.

The proper initialization should ensure that the list is always correct,
but just in case we check its size and return a qCritical alert.