Added dialog to select initial contour value
Now, when a contour is being initialized, a dialog box will appear
asking the user to confirm the initial contour value. The new default
value for the contour is 2/3 of the range instead of 1/2.

If the user initializes the contour by double-clicking on the histogram,
the value at which they double-clicked will be the default value.

The user may also check "Don't ask again" if they do not wish to
see this dialog box again.